Extract more methods to view for presenter and create observer class
