+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationContract.java
@@ -1,5 +1,7 @@
 package com.mapbox.services.android.navigation.ui.v5;
 
+import android.location.Location;
+
 import com.mapbox.api.directions.v5.models.DirectionsRoute;
 import com.mapbox.geojson.Point;
 
@@ -26,5 +28,11 @@ public interface NavigationContract {
     void finishNavigationView();
 
     void takeScreenshot();
+
+    void startCamera(DirectionsRoute directionsRoute);
+
+    void resumeCamera(Location location);
+
+    void updateLocationLayer(Location location);
   }
 }
