+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationViewModelObserver.java
@@ -1,0 +1,10 @@
+package com.mapbox.services.android.navigation.ui.v5;
+
+import android.content.Context;
+
+public class NavigationViewModelObserver {
+
+  public NavigationViewModelObserver(Context context, NavigationPresenter presenter, NavigationViewListener listener) {
+
+  }
+}
